requirements.txt
flask>=3.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
gunicorn>=22.0
